requirements.txt

fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy>=2.0.25,<3.0.0
asyncpg>=0.31.0,<1.0.0
psycopg2-binary>=2.9.9,<3.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0
greenlet>=3.0.0
